some updates in program

+++ models/mealModel.js
@@ -11,10 +11,10 @@ const Food = sequelize.define(
         serving: { type: DataTypes.STRING, allowNull: false },
         title: { type: DataTypes.STRING, allowNull: false },
         image_url: { type: DataTypes.TEXT },
-        fat: { type: DataTypes.INTEGER },
-        cals: { type: DataTypes.INTEGER },
-        carbs: { type: DataTypes.INTEGER },
-        protein: { type: DataTypes.INTEGER },
+        fat: { type: DataTypes.FLOAT },
+        cals: { type: DataTypes.FLOAT },
+        carbs: { type: DataTypes.FLOAT },
+        protein: { type: DataTypes.FLOAT },
     },
     { timestamps: true, createdAt: true, updatedAt: false }
 )
